+++ src/design/data/icednano/tests/mod.rs
@@ -17,6 +17,7 @@ ENSnano, a 3d graphical application for DNA nanostructures.
 */
 
 use super::*;
+use std::fmt::Write;
 
 #[test]
 fn sanitize_with_insertions() {
@@ -84,6 +85,47 @@ fn sanitize_domains_scadnano() {
     );
 }
 
+#[test]
+fn scadnano_import_one_loopout() {
+    let input = r##" {
+  "version": "0.15.0",
+  "grid": "square",
+  "helices": [
+    {"grid_position": [0, 0]},
+    {"grid_position": [0, 1]}
+  ],
+  "strands": [
+    {
+      "circular": true,
+      "color": "#57bb00",
+      "domains": [
+        {"helix": 0, "forward": true, "start": 8, "end": 16},
+        {"loopout": 5},
+        {"helix": 1, "forward": false, "start": 8, "end": 16}
+      ]
+    }
+  ]
+      }"##;
+    let scadnano_design: super::super::scadnano::ScadnanoDesign =
+        serde_json::from_str(&input).expect("Failed to parse scadnano input");
+    let ensnano_design =
+        Design::from_scadnano(&scadnano_design).expect("Could not convert to ensnano");
+    assert_eq!(ensnano_design.strands.len(), 1);
+    let strand = ensnano_design.strands.values().next().unwrap();
+    assert_good_strand(&strand, "[H0: 8 -> 15] [@5] [H1: 8 <- 15]")
+}
+
+fn assert_good_strand<S: std::ops::Deref<Target = str>>(strand: &Strand, objective: S) {
+    use regex::Regex;
+    let re = Regex::new(r#"\[[^\]]*\]"#).unwrap();
+    let formated_strand = strand.formated_domains();
+    let left = re.find_iter(&formated_strand);
+    let right = re.find_iter(&objective);
+    for (a, b) in left.zip(right) {
+        assert_eq!(a.as_str(), b.as_str());
+    }
+}
+
 fn assert_sane_domains_non_cyclic(dom: &[Domain]) {
     let mut prev_insertion = false;
     for d in dom.iter() {
@@ -111,12 +153,10 @@ fn assert_sane_domains_cyclic(dom: &[Domain]) {
 
 #[test]
 fn correct_sanetization() {
-    let strand = strand_with_insertion();
+    let mut strand = strand_with_insertion();
     let sane_domains = sanitize_domains(&strand.domains, false);
-    assert_eq!(
-        sane_domains.iter().map(|d| d.length()).collect::<Vec<_>>(),
-        vec![4, 8, 4, 5, 8]
-    );
+    strand.domains = sane_domains;
+    assert_good_strand(&strand, formated_sane_strand_with_insertion());
 }
 
 #[test]
@@ -143,6 +183,10 @@ fn correct_sanetization_cyclic_pathological() {
         sane_domains.iter().map(|d| d.length()).collect::<Vec<_>>(),
         vec![4, 8, 4, 5, 8, add_prime5 + add_prime3]
     );
+    strand.domains = sane_domains;
+    let mut expected = String::from(formated_sane_strand_with_insertion());
+    write!(&mut expected, "[@{}]", add_prime5 + add_prime3).unwrap();
+    assert_good_strand(&strand, expected);
 }
 
 #[test]
@@ -517,8 +561,87 @@ fn correct_junction_cyclic_pathological() {
     );
 }
 
+#[test]
+fn test_insertion_left_to_right() {
+    let mut strand = strand_with_insertion();
+    let domains = sanitize_domains(&strand.domains, strand.cyclic);
+    strand.domains = domains;
+    let first_insertion = 1552;
+    strand.add_insertion_at_nucl(
+        &Nucl {
+            helix: 1,
+            position: 6,
+            forward: true,
+        },
+        first_insertion,
+    );
+
+    let second_insertion = 172634;
+    strand.add_insertion_at_nucl(
+        &Nucl {
+            helix: 2,
+            position: 5,
+            forward: false,
+        },
+        second_insertion,
+    );
+
+    let objective = format!(
+        "[H1: 0 -> 3] [@8] [H1: 4 -> 6] [@{}] [H1: 7 -> 7] [@5] [H2: 5 <- 7] [@{}] [H2: 0 <- 4]",
+        first_insertion, second_insertion
+    );
+    assert_good_strand(&strand, objective);
+}
+
+#[test]
+fn test_insertion_right_to_left() {
+    let mut strand = strand_with_insertion();
+    let domains = sanitize_domains(&strand.domains, strand.cyclic);
+    strand.domains = domains;
+    let first_insertion = 1552;
+
+    let second_insertion = 172634;
+    strand.add_insertion_at_nucl(
+        &Nucl {
+            helix: 2,
+            position: 5,
+            forward: false,
+        },
+        second_insertion,
+    );
+    strand.add_insertion_at_nucl(
+        &Nucl {
+            helix: 1,
+            position: 6,
+            forward: true,
+        },
+        first_insertion,
+    );
+
+    let objective = format!(
+        "[H1: 0 -> 3] [@8] [H1: 4 -> 6] [@{}] [H1: 7 -> 7] [@5] [H2: 5 <- 7] [@{}] [H2: 0 <- 4]",
+        first_insertion, second_insertion
+    );
+    assert_good_strand(&strand, objective);
+}
+
+/// A strand whose inital topology is [H1: 0 -> 3] [@5] [@3] [H1: 4 -> 7] [@5] [H2: 0 <- 7]
 fn strand_with_insertion() -> Strand {
     let strand_str = include_str!("./strand_with_insertion.json");
     let strand: Strand = serde_json::from_str(&strand_str).expect("Could not parse strand");
     strand
+}
+
+fn formated_strand_with_insertion() -> &'static str {
+    "[H1: 0 -> 3] [@5] [@3] [H1: 4 -> 7] [@5] [H2: 0 <- 7]"
+}
+
+fn formated_sane_strand_with_insertion() -> &'static str {
+    "[H1: 0 -> 3] [@8] [H1: 4 -> 7] [@5] [H2: 0 <- 7]"
+}
+
+#[test]
+fn check_formated_strand_with_insertion() {
+    let strand = strand_with_insertion();
+    assert_good_strand(&strand, formated_strand_with_insertion())
 }